requirements.txt
anyio==4.6.2.post1
fastapi==0.115.5
motor==3.6.0
passlib==1.7.4
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==1.7.1
pymongo==4.9.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.2.0
uvicorn==0.32.0